+++ app/src/main/java/com/example/tudeeapp/di/ViewModelModule.kt
@@ -11,7 +11,7 @@ import com.example.tudeeapp.presentation.screen.home.HomeViewModel
 import com.example.tudeeapp.presentation.screen.onBoarding.OnBoardingViewModel
 import com.example.tudeeapp.presentation.screen.task.TasksViewModel
 import com.example.tudeeapp.presentation.screen.taskDetails.TaskDetailsViewModel
-import com.example.tudeeapp.presentation.screen.taskManagement.TaskManagementViewModel
+import com.example.tudeeapp.presentation.screen.taskForm.TaskFormViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -22,7 +22,7 @@ val viewModelModule = module {
     viewModelOf(::TaskDetailsViewModel)
     viewModelOf(::CategoriesViewModel)
     viewModelOf(::OnBoardingViewModel)
-    viewModelOf(::TaskManagementViewModel)
+    viewModelOf(::TaskFormViewModel)
     viewModelOf(::CategoryDetailsViewModel)
     viewModelOf(::CategoryFormViewModel)
     viewModelOf(::TasksViewModel)
